restructured several components
