+++ src/components/spaceships/cards/StaticInfoCard.module.css
@@ -1,17 +1,3 @@
-.ship {
-  margin: 5% auto;
-  max-width: 960px;
-  display: flex;
-  border-radius: 5%;
-  overflow: auto;
-}
-
-.ship img {
-  max-width: 640px;
-  opacity: 0.8;
-  border-radius: 5% 0 0 5%;
-}
-
 .ship-info {
   color: #d4d8de;
   text-shadow: 1px 1px 2px #080808;
@@ -36,7 +22,7 @@
   margin: 0 0 2rem;
 }
 
-.ship article {
+.ship-info article {
   padding: 2rem;
   background: rgb(255, 255, 255, 0.1);
 }
@@ -59,7 +45,7 @@
 }
 
 @media screen and (max-width: 900px) {
-  .ship, .ship img, .ship-info {
+  .ship img, .ship-info {
     border-radius: 0;
   }
 }
