+++ src/components/spaceships/list/ShipItem.module.css
