+++ src/components/spaceships/list/ShipItem.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from 'react-router-dom';
-import SpaceShip from '../../models/SpaceShip';
+import SpaceShip from '../../../models/SpaceShip';
 import { motion } from 'framer-motion';
 import css from './ShipItem.module.css';
-import FontMotion from '../UI/FontMotion';
+import FontMotion from '../../UI/FontMotion';
 
 interface ShipItemProps extends SpaceShip {
   className: string;
