+++ src/pages/spaceships/ShipID.tsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { spaceships } from '../../assets/spaceships/spaceships';
-import ShipInfo from '../../components/spaceships/ShipInfo';
+import ShipDetails from '../../components/spaceships/ShipDetails';
 
 export default function ShipIDPage() {
   const { id } = useParams();
@@ -12,6 +12,6 @@ export default function ShipIDPage() {
     throw new Error('No Ship Found');
   }
 
-  return <ShipInfo {...spaceship} />;
+  return <ShipDetails {...spaceship} />;
   // key will retrigger animations for this component each time we click on a new ship
 }
